portfoy-ai-studio/requirements.txt
tefas-crawler>=0.5.0
psycopg2-binary>=2.9.9
requests>=2.32.0
pandas>=2.0.0
